Refactor sign/verify to their own funcs that can be called without status outputs.
